switch to devbox github action
Since Node and pnpm are build-time development dependencies for most packages, there's no reason to deal with seemingly-random failures caused by Node/pnpm version mismatches in github actions for this project. Let's just use the same version that is in devbox so that my local development environment perfectly matches what runs in github actions.

This is at the expense of using a test matrix for the Node version and making sure we build against lts and current.