upgrade satori to v0.12.2
https://github.com/vercel/satori/commit/d1dfccef303137cfcedf6a9cdf79624b257ae3e0 seems to have fixed an issue with Satori-generated content bleeding out of the svg bounding box. I'm removing some of the hacks we had in place to account for this.

To be precise: it looks like Satori can still produce svg children (e.g. letters) that extend out of the svg bounding box, but if we compute the real width/height as the rightmost pixel minus the leftmost pixel, I still get the intended width, so we don't need to _account_ for this bleed anymore.